+++ jest.config.js
@@ -1,10 +1,16 @@
 const files = `<rootDir>/test/unit/modules/*${
   process.env.TEST_ENV ? `-${process.env.TEST_ENV}-` : ``
-}*browser-spec.js`;
+}*.js`;
 
 module.exports = {
   verbose: true,
   bail: true,
   testMatch: [files],
-  testEnvironment: process.env.TEST_ENV || 'jsdom'
+  testEnvironment: process.env.TEST_ENV || 'jsdom',
+  transform: {
+      "^.+\\.js$": "babel-jest",
+  },
+  transformIgnorePatterns: [
+    "node_modules/(?!promise-polyfill|whatwg-fetch)/"
+  ]
 };